pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.8.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
]
